refactor: moved pino logger
