+++ src/shared/libs/logger/pino/index.ts
@@ -1,0 +1,1 @@
+export { PinoLogger } from './pino-logger.js';
